requirements.txt
python-dateutil
pyspark
warcio
langdetect
newspaper3k
